pyproject.toml
[project]
name = "autobahn"
dynamic = ["version"]
description = "agent orchestration superhighway"
authors = [
    { name = "David Lee", email = "dav@davlee.com" }
]
readme = "README.md"
license = { text = "FSL-1.1-ALv2" }
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0,<3.0",
    "pyyaml>=6.0.3",
    "tomlkit>=0.13.0",
]
keywords = ["agent", "orchestration", "session", "lifecycle"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.urls]
Homepage = "https://github.com/davidlee/autobahn"
Repository = "https://github.com/davidlee/autobahn"

#
# BUILD MANIFEST
#

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "VERSION"
pattern = "(?P<version>.+)"

[tool.hatch.build.targets.wheel]
packages = ["autobahn"]
exclude = [
    ".direnv/",
    ".envrc",
    ".venv/",
    "flake.nix",
    "flake.lock",
    ".uv-",
    ".uv-cache/",
    ".pytest_cache/",
    "**/__pycache__/",
    "**/*.pyc",
    "**/*.pyo",
]

[tool.hatch.build.targets.sdist]
packages = ["autobahn"]
exclude = [
    ".direnv/",
    ".envrc",
    ".venv/",
    "flake.nix",
    "flake.lock",
    ".uv-",
    ".uv-cache/",
    ".pytest_cache/",
    "**/__pycache__/",
    "**/*.pyc",
    "**/*.pyo",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.0",
    "ruff>=0.14.0",
    "pylint>=4.0.2",
    "pylint-per-file-ignores>=3.1.0",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra"
asyncio_mode = "auto"
testpaths = [
    "tests",
]

[tool.ruff]
line-length = 88
indent-width = 2
target-version = "py312"

[tool.ruff.lint]
extend-select = [
    "F",    # Pyflakes
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "A",    # flake8-builtins
    "C4",   # flake8-comprehensions
    "ICN",  # import conventions
    "RET",  # return values
    "PTH",  # Use pathlib instead of os.path
    "COM",  # enforce trailing comma rules
    "TID",  # import practices
    "FA",   # from __future__ import annotations
    "SIM",  # simplify
    "PL",   # pylint (selective)
]
ignore = [
    "T201",    # print
    "S101",    # assert
    "PLR0911", # too-many-return-statements
    "PLR0912", # too-many-branches
    "PLR0913", # too-many-arguments
    "PLR0915", # too-many-statements
    "C901",    # complex-structure
    "FBT001", "FBT002",  # Boolean args
    "PLR2004", # Magic values
    "D",       # pydocstyle
    "PTH123",  # Allow open()
    "COM812",  # trailing comma (formatter conflict)
    "UP046",   # PEP 695 type params (pydantic Generic compat)
    "UP047",   # PEP 695 type params (pydantic Generic compat)
]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"*test*.py" = [
    "S101",    # asserts in tests
    "ARG",     # unused fixture args
    "FBT",     # boolean positional args
    "PLR2004", # magic values
    "S311",    # pseudo-random
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

#
# PyLint
#

[tool.pylint.main]
fail-under = 9.70
reports = true
output-format = "colorized"
indent-string = "  "
load-plugins = [
    "pylint_per_file_ignores",
    "pylint.extensions.docstyle",
    "pylint.extensions.docparams",
    "pylint.extensions.mccabe",
]

[tool.pylint.parameter_documentation]
default-docstring-type = "google"

[tool.pylint.format]
indent-string = "  "
expected-line-ending-format = "LF"

[tool.pylint.'messages control']
per-file-ignores = [
    "**/*_test.py:protected-access,missing-function-docstring",
    "**/test_*.py:protected-access,missing-function-docstring",
]

[tool.pylint.design]
exclude-too-few-public-methods = []
ignored-parents = []
max-args = 5
max-attributes = 7
max-bool-expr = 5
max-branches = 12
max-locals = 15
max-parents = 7
max-positional-arguments = 5
max-public-methods = 20
max-returns = 6
max-statements = 50
min-public-methods = 2
